services/auth-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator